Remove InitialArticleZip workflow, version reason decider.
