+++ newFileWorkflows.yaml
@@ -1,7 +1,7 @@
 ArticleZip:
   bucket_name_pattern: '.*elife-production-final$'
   file_name_pattern: '.*\.zip'
-  starter_name: 'InitialArticleZip'
+  starter_name: 'IngestArticleZip'
 
 SilentCorrectionsArticleZip:
   bucket_name_pattern: '.*elife-silent-corrections$'
