+++ starter/starter_IngestArticleZip.py
@@ -1,6 +1,6 @@
 import boto.swf
 import json
-from optparse import OptionParser
+from S3utility.s3_notification_info import S3NotificationInfo
 import starter.starter_helper as helper
 from starter.starter_helper import NullRequiredDataException
 
@@ -13,17 +13,18 @@ class starter_IngestArticleZip():
     def __init__(self):
         self.const_name = "IngestArticleZip"
         
-    def start(self, settings, run, article_id, version_reason=None, scheduled_publication_date=None):
+    def start(self, settings, run, info):
 
         # Log
         logger = helper.get_starter_logger(settings.setLevel, helper.get_starter_identity(self.const_name))
 
-        input = {
-                    "run": run,
-                    "article_id": article_id,
-                    "version_reason": version_reason,
-                    "scheduled_publication_date": scheduled_publication_date
-        }
+        if hasattr(info, 'file_name') == False or info.file_name is None:
+            raise NullRequiredDataException("filename is Null. Did not get a filename.")
+
+        input = S3NotificationInfo.to_dict(info)
+        input['run'] = run
+        input['version_lookup_function'] = "article_next_version"
+        input['run_type'] = "initial-article"
 
         workflow_id, \
         workflow_name, \
@@ -31,7 +32,7 @@ class starter_IngestArticleZip():
         child_policy, \
         execution_start_to_close_timeout, \
         workflow_input = helper.set_workflow_information(self.const_name, "1", None, input,
-                                                         "ingest-%s" % article_id,
+                                                         info.file_name.replace('/', '_'),
                                                          start_to_close_timeout=str(60 * 60 * 1))
 
         # Simple connect
@@ -53,28 +54,3 @@ class starter_IngestArticleZip():
             message = 'SWFWorkflowExecutionAlreadyStartedError: ' \
                       'There is already a running workflow with ID %s' % workflow_id
             logger.info(message)
-
-
-if __name__ == "__main__":
-
-    doi_id = None
-
-    # Add options
-    parser = OptionParser()
-    parser.add_option("-e", "--env", default="dev", action="store", type="string", dest="env",
-                      help="set the environment to run, either dev or live")
-    parser.add_option("-f", "--filename", default=None, action="store", type="string", dest="filename",
-                      help="specify the DOI id the article to process")
-
-    (options, args) = parser.parse_args()
-    if options.env:
-        ENV = options.env
-    if options.filename:
-        filename = options.filename
-
-    import settings as settingsLib
-    settings = settingsLib.get_settings(ENV)
-
-    o = starter_IngestArticleZip()
-
-    o.start(settings=settings,)
