+++ tests/starter/test_starter_ingest_article_zip.py
@@ -8,29 +8,20 @@ from mock import patch
 from tests.classes_mock import FakeBotoConnection
 
 
+RUN_EXAMPLE = u'1ee54f9a-cb28-4c8e-8232-4b317cf4beda'
+
+
 class TestStarterIngestArticleZip(unittest.TestCase):
     def setUp(self):
-        self.stater_ingest_article_zip = starter_IngestArticleZip()
+        self.starter = starter_IngestArticleZip()
+
+    def test_ingest_article_zip_starter_no_article(self):
+        self.assertRaises(NullRequiredDataException, self.starter.start,
+                          settings=settings_mock, run=RUN_EXAMPLE, info=test_data.data_error_lax)
 
     @patch('starter.starter_helper.get_starter_logger')
     @patch('boto.swf.layer1.Layer1')
     def test_ingest_article_zip_starter_(self, fake_boto_conn, fake_logger):
         fake_boto_conn.return_value = FakeBotoConnection()
-        self.stater_ingest_article_zip.start(settings=settings_mock, run=test_data.ingest_article_zip_data['run'],
-                                             article_id=test_data.ingest_article_zip_data['article_id'],
-                                             version_reason=test_data.ingest_article_zip_data['version_reason'],
-                                             scheduled_publication_date=test_data.ingest_article_zip_data[
-                                                 'scheduled_publication_date']
-                                             )
-
-    @patch('starter.starter_helper.get_starter_logger')
-    @patch('boto.swf.layer1.Layer1')
-    def test_ingest_article_zip_starter_novr_(self, fake_boto_conn, fake_logger):
-        fake_boto_conn.return_value = FakeBotoConnection()
-        self.stater_ingest_article_zip.start(settings=settings_mock, run=test_data.ingest_article_zip_no_vr_data['run'],
-                                             article_id=test_data.ingest_article_zip_no_vr_data['article_id']
-                                             )
-
-
-if __name__ == '__main__':
-    unittest.main()
+        self.starter.start(settings=settings_mock, run=RUN_EXAMPLE,
+                                             info=S3NotificationInfo.from_dict(test_data.ingest_article_zip_data))
