--- tests/starter/test_starter_initial_article_zip.py
@@ -1,29 +1,0 @@
-import unittest
-from starter.starter_InitialArticleZip import starter_InitialArticleZip
-from starter.starter_helper import NullRequiredDataException
-from S3utility.s3_notification_info import S3NotificationInfo
-import tests.settings_mock as settings_mock
-import tests.test_data as test_data
-from mock import patch
-from tests.classes_mock import FakeBotoConnection
-
-run_example = u'1ee54f9a-cb28-4c8e-8232-4b317cf4beda'
-
-class TestStarterInitialArticleZip(unittest.TestCase):
-    def setUp(self):
-        self.stater_initial_article_zip = starter_InitialArticleZip()
-
-    def test_initial_article_zip_starter_no_article(self):
-        self.assertRaises(NullRequiredDataException, self.stater_initial_article_zip.start,
-                          settings=settings_mock, run=run_example, info=test_data.data_error_lax)
-
-    @patch('starter.starter_helper.get_starter_logger')
-    @patch('boto.swf.layer1.Layer1')
-    def test_initial_article_zip_starter_(self, fake_boto_conn, fake_logger):
-        fake_boto_conn.return_value = FakeBotoConnection()
-        self.stater_initial_article_zip.start(settings=settings_mock, run=run_example,
-                                             info=S3NotificationInfo.from_dict(test_data.initial_article_zip_data))
-
-
-if __name__ == '__main__':
-    unittest.main()
