--- workflow/workflow_InitialArticleZip.py
@@ -1,48 +1,0 @@
-from workflow.objects import Workflow
-from workflow.helper import define_workflow_step, define_workflow_step_medium
-
-
-class workflow_InitialArticleZip(Workflow):
-    def __init__(self, settings, logger, conn=None, token=None, decision=None,
-                 maximum_page_size=100):
-        super(workflow_InitialArticleZip, self).__init__(
-            settings, logger, conn, token, decision, maximum_page_size)
-
-        # SWF Defaults
-        self.name = "InitialArticleZip"
-        self.version = "1"
-        self.description = "Initial workflow for article processing"
-        self.default_execution_start_to_close_timeout = 60 * 5
-        self.default_task_start_to_close_timeout = 30
-
-        # Get the input from the JSON decision response
-        data = self.get_input()
-
-        # JSON format workflow definition, for now - may be from common YAML definition
-        workflow_definition = {
-            "name": self.name,
-            "version": self.version,
-            "task_list": self.settings.default_task_list,
-            "input": data,
-
-            "start":
-                {
-                    "requirements": None
-                },
-
-            "steps":
-                [
-                    define_workflow_step("PingWorker", data),
-                    define_workflow_step_medium("VersionLookup", data),
-                    define_workflow_step_medium("ExpandArticle", data),
-                    define_workflow_step_medium("SendDashboardProperties", data),
-                    define_workflow_step_medium("VersionReasonDecider", data),
-                ],
-
-            "finish":
-                {
-                    "requirements": None
-                }
-        }
-
-        self.load_definition(workflow_definition)
